requirements.txt
deepl
flask
gunicorn
Pillow
pyocr
